Order Controller - moved Exception from Controller to Service layer

+++ src/main/java/com/kodilla/ecommercee/service/OrderService.java
@@ -2,7 +2,9 @@ package com.kodilla.ecommercee.service;
 
 import com.kodilla.ecommercee.domain.Order;
 import com.kodilla.ecommercee.repository.OrderRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -25,10 +27,16 @@ public class OrderService {
     }
 
     public Optional<Order> getOrder(Long orderId){
+        if(!orderRepository.existsById(orderId)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "order not found");
+        }
         return orderRepository.findById(orderId);
     }
 
     public void delete(Long orderId){
+        if(!orderRepository.existsById(orderId)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "order not found");
+        }
         orderRepository.deleteById(orderId);
     }
 }
